Reduce the quantity of heap allocations via the introduction of a back buffer
Reallocating the game state at each iteration of the game is expensive.
Instead, to lessen the number of allocations, double buffering is preferred.
Throughout the entirety of any game, only two allocations are necessary to
compute the game state: one for the buffer tracking the current game state
(the front buffer), and one for the buffer tracking the succeeding game
state (the back buffer).