Create parameter to enable background color
The asciigol program will not alter the background and foreground colors of
the terminal by default. However, a new parameter has been added to allow the
background color to change. The background color type can be "none" (default),
"light" (white background, black foreground), or "dark" (black background,
white foreground). In the event that the live and dead characters are the same
and the background color is either "light" or "dark", then the colors will
alternate between live and dead cells. For example, if the background color
type is "dark" and both live and dead characters are the same, the live cell
will be printed with a black background and white foreground, whereas the dead
cell will be printed with a white background and black foreground.